// src/types/api.ts
export interface NobelPrizePerLaureate {
  /** @example 1989 */
  awardYear?: number;
  category?: Translation;
  categoryFullName?: Translation;
  /** @default "1" */
  sortOrder?: "1" | "2" | "3";
  /** @default "1" */
  portion?: "1" | "1/2" | "1/3" | "1/4";
  /**
   * @format date
   * @example "1989-04-25"
   */
  dateAwarded?: string;
  /** @example "received" */
  prizeStatus?: "received" | "declined" | "restricted";
  motivation?: Translation;
  /**
   * Prize amount given at the year of the award
   * @example 1000000
   */
  prizeAmount?: number;
  /**
   * Prize amount adjusted according to Index number yearly average
   * @example 1000000
   */
  prizeAmountAdjusted?: number;
  affiliations?: Entity[];
  residences?: Residence[];
  links?: ItemLinks[];
}

/** Short information about the Laureate */
export interface LaureateBasic {
  /**
   * @min 1
   * @example 34
   */
  id?: number;
  name?: Translation;
  /** @default "1" */
  portion?: "1" | "1/2" | "1/3" | "1/4";
  /** @default "1" */
  sortOrder?: "1" | "2" | "3";
  motivation?: Translation;
  links?: ItemLinks[];
}

/** Laureate's full information and their respective awarded Prizes */
export interface Laureate {
  /**
   * @min 1
   * @example 456
   */
  id?: number;

  // Person-specific fields
  knownName?: Translation;
  givenName?: Translation;
  familyName?: Translation;
  fullName?: Translation;
  /**
   * Short slug version of family name used in filenames
   * @example "filename"
   */
  filename?: string;
  /**
   * Only for some Laureates in Literature category
   * @example "Penname"
   */
  penname?: string;
  /** @example "female" */
  gender?: "female" | "male";
  birth?: Event;
  death?: Event;

  // Organization-specific fields
  orgName?: Translation;
  /** @example "Native name" */
  nativeName?: string;
  /** @example "Acronym" */
  acronym?: string;
  founded?: Event;
  dissolution?: Event;
  /** City and county at the time, citynow and countrynow are as up-to-date names. Continent as geonames.org standard. */
  headquarters?: Location;

  // Other fields
  wikipedia?: LaureateWikipedia;
  wikidata?: LaureateWikidata;
  sameAs?: Urls[];
  links?: ItemLinks[];
  nobelPrizes?: NobelPrizePerLaureate[];
}
// ...
export interface Location {
  city?: Translation;
  country?: Translation;
  cityNow?: TranslationWithSameas;
  countryNow?: TranslationWithSameas;
  continent?: Translation;
  locationString?: Translation;
}

export interface Event {
  /**
   * @format date
   * @example "1989-04-25"
   */
  date?: string;
  /** City and county at the time, citynow and countrynow are as up to date names. Continent as geonames.org standard. LocationString is the combination of city, country, citynow and country now. */
  place?: Location;
}
// ...
export interface ItemLinks {
  rel?: string;
  href?: string;
  action?: string;
  types?: string;
}
// ...
export interface Entity {
  name?: Translation;
  nameNow?: Translation;
  /** @example "Le romenic universitet" */
  nativeName?: string;
  city?: Translation;
  country?: Translation;
  cityNow?: TranslationWithSameas;
  countryNow?: TranslationWithSameas;
  locationString?: Translation;
}

export interface Residence {
  city?: Translation;
  country?: Translation;
  cityNow?: TranslationWithSameas;
  countryNow?: TranslationWithSameas;
  locationString?: Translation;
}

export interface Translation {
  /** @example "English" */
  en?: string;
  /** @example "Svenska" */
  se?: string;
  /** @example "Norsk" */
  false?: string;
}

export interface TranslationWithSameas {
  /** @example "English" */
  en?: string;
  /** @example "Svenska" */
  se?: string;
  /** @example "Norsk" */
  false?: string;
  sameAs?: Urls[];
}

/** @example "https://" */
export type Urls = string;

export interface Error {
  /** @example "404" */
  code: string;
  /** @example "There is not Laureate born that date" */
  message: string;
}

export interface LaureateWikipedia {
  /**
   * Wikipedia url slug
   * @example "slug"
   */
  slug?: string;
  /**
   * URL to english wikipedia page
   * @example "https://en.wikipedia.org/wiki/"
   */
  english?: string;
}

export interface LaureateWikidata {
  /**
   * Wikidata id
   * @example "id"
   */
  id?: string;
  /**
   * URL to wikidata entry
   * @example "https://www.wikidata.org/wiki/"
   */
  url?: string;
}